sprint 12: sse 구조 변경, 파일 업로드 상태 변경 이벤트 전송
