+++ src/main/java/com/spirnt/mission/discodeit/controller/SseController.java
@@ -1,7 +1,7 @@
 package com.spirnt.mission.discodeit.controller;
 
 import com.spirnt.mission.discodeit.security.CustomUserDetails;
-import com.spirnt.mission.discodeit.service.NotificationService;
+import com.spirnt.mission.discodeit.sse.SseEmitterManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -18,13 +18,13 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 @RequestMapping("/api/sse")
 public class SseController {
 
-    private final NotificationService notificationService;
+    private final SseEmitterManager emitterManager;
 
     @GetMapping("")
     public ResponseEntity<SseEmitter> sseEmitter(
         @AuthenticationPrincipal CustomUserDetails customUserDetails,
         @RequestHeader(value = "Last-Event-ID", required = false) String lastEventId) {
-        SseEmitter sseEmitter = notificationService.subscribe(customUserDetails.getUser().getId(),
+        SseEmitter sseEmitter = emitterManager.subscribe(customUserDetails.getUser().getId(),
             lastEventId);
         return ResponseEntity.ok(sseEmitter);
     }
